% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{autoplot.locfdr}
\alias{autoplot.locfdr}
\title{Plot a locfdr Object with ggplot2}
\usage{
autoplot.locfdr(object, ...)
}
\arguments{
\item{object}{An object of class `"locfdr"`.}

\item{...}{Additional arguments (ignored).}
}
\value{
A [ggplot2::ggplot] object.
}
\description{
Creates a ggplot2 visualization of the local false discovery rate
estimation, showing the histogram of z-values with the fitted mixture
density and null subdensity overlay.
}
