% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hivdata}
\alias{hivdata}
\title{HIV Gene Expression Data}
\format{
A numeric vector containing 7680 z-values.
}
\usage{
hivdata
}
\description{
The data comprises 7680 z-values, each relating to a two-sample
t-test.  The test compares gene expression values for 4 HIV patients
with values for 4 normal subjects; the t-score `T[i]` for gene i has
been transformed to a normal scale, `z[i] = qnorm(pt(T[i], df=6))`,
so that the `z[i]`'s theoretically would have a standard \eqn{N(0,1)}
distribution under the null hypothesis.
}
\examples{
data(hivdata)
hist(hivdata, breaks = 100)
}
\references{
van't Wout, et al., Cellular gene expression upon human
  immunodeficiency virus type 1 infection of CD4+-T-Cell lines,
  \emph{Journal of Virology} 77, 1392-1402.
}
\keyword{datasets}
