% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfdrsim}
\alias{lfdrsim}
\title{Simulated Data for locfdr}
\format{
A data frame with 2000 rows and 2 columns:
\describe{
  \item{mu}{the true score for each gene}
  \item{zex}{the observed z-value}
}
}
\usage{
lfdrsim
}
\description{
A simulated dataset that involves 2000 "genes", each of which has
yielded a test statistic "zex", with \eqn{zex[i] \sim N(mu[i], 1)}
(independently for \eqn{i = 1, 2, \ldots, 2000}).  Of the 2000
\eqn{\mu_i} values, 1800 (90\%) are zero (null cases) and the
remaining 200 (10\%) are drawn from a \eqn{N(3, 1)} distribution
(non-null cases).
}
\examples{
data(lfdrsim)
hist(lfdrsim$zex, breaks = 80)
}
\keyword{datasets}
