% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.locfdr}
\alias{print.locfdr}
\title{Print a locfdr Object}
\usage{
\method{print}{locfdr}(x, ...)
}
\arguments{
\item{x}{An object of class `"locfdr"`.}

\item{...}{Additional arguments (ignored).}
}
\value{
`x`, invisibly.
}
\description{
Print a locfdr Object
}
