% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.locfdr}
\alias{summary.locfdr}
\title{Summarize a locfdr Object}
\usage{
\method{summary}{locfdr}(object, ...)
}
\arguments{
\item{object}{An object of class `"locfdr"`.}

\item{...}{Additional arguments (ignored).}
}
\value{
`object`, invisibly.
}
\description{
Summarize a locfdr Object
}
