\name{lfdrsim}
\docType{data}
\alias{lfdrsim}
\title{Simulated data set for locfdr}
\description{A simulated dataset that involves 2000 "genes", each of
  which has yielded a test statistic "zex", with
  \eqn{zex[i] ~ N(mu[i],1)} (independently for \eqn{i=1,2,...2000.})
  The data comprises 2000 \eqn{\mu_i}{mu} values and 2000
  \eqn{z}-values.
}
\format{A matrix of 2000 rows and 2 columns containing mu and the
  z-score values (\code{zex})}
\usage{data(lfdrsim)}
\keyword{datasets}
